feat: default to orpheus_cpp and isolate remote backend
