+++ Morpheus_Client/admin/tts.html
@@ -157,7 +157,7 @@
                       <label for="model" class="block text-sm font-medium text-white mb-1">Model</label>
                       <div class="relative">
                         <select id="model" name="model" class="block w-full rounded-md bg-dark-700 border-dark-600 text-white shadow-sm focus:border-primary-500 focus:ring-primary-500 focus:ring-offset-dark-800 focus:outline-none outline-none sm:text-sm pl-3 pr-10 py-2 appearance-none">
-                          <option value="orpheus" selected>Orpheus 3B (0.1)</option>
+                          <option value="orpheus_cpp" selected>Orpheus 3B (0.1)</option>
                         </select>
                         <div class="pointer-events-none absolute inset-y-0 right-0 flex items-center px-2 text-purple-300">
                           <svg class="h-5 w-5" fill="currentColor" viewBox="0 0 20 20">
